[miner] Kato -> Coordinator
