+++ src/Chainweb/Miner/RestAPI/Server.hs
@@ -32,8 +32,8 @@ import Chainweb.BlockHeader (BlockHeader(..), decodeBlockHeaderWithoutHash)
 import Chainweb.Chainweb.MinerResources (MinerResources(..))
 import Chainweb.CutDB (CutDb)
 import Chainweb.Logger (Logger, logFunction)
+import Chainweb.Miner.Coordinator (MiningState(..), publish)
 import Chainweb.Miner.Core (HeaderBytes(..))
-import Chainweb.Miner.Kato (MiningState(..), publish)
 import Chainweb.Miner.RestAPI (MiningResultApi)
 import Chainweb.RestAPI.Utils (SomeServer(..))
 import Chainweb.Utils (runGet)
